+++ app/blueprints/base_blueprint.py
@@ -14,13 +14,13 @@ class BaseBlueprint:
 		
 		''' Register All App Blue Prints Here '''
 		
-		from app.blueprints.meal_blueprint import meal_blueprint
+		from app.blueprints.meal_item_blueprint import meal_item_blueprint
 		from app.blueprints.vendor_blueprint import vendor_blueprint, engagement_blueprint, rating_blueprint
 		from app.blueprints.location_blueprint import location_blueprint
 		from app.blueprints.role_blueprint import role_blueprint
 		from app.blueprints.menu_blueprint import menu_blueprint
 		
-		self.app.register_blueprint(meal_blueprint)
+		self.app.register_blueprint(meal_item_blueprint)
 		self.app.register_blueprint(vendor_blueprint)
 		self.app.register_blueprint(engagement_blueprint)
 		self.app.register_blueprint(rating_blueprint)
