+++ app/controllers/menu_controller.py
@@ -1,6 +1,6 @@
 from app.controllers.base_controller import BaseController
 from app.repositories.menu_repo import MenuRepo
-from app.repositories.meal_repo import MealItemRepo
+from app.repositories.meal_item_repo import MealItemRepo
 
 
 class MenuController(BaseController):
