+++ app/repositories/menu_repo.py
@@ -1,6 +1,6 @@
 from app.repositories.base_repo import BaseRepo
 from app.models.menu import Menu
-from app.repositories.meal_repo import MealItemRepo
+from app.repositories.meal_item_repo import MealItemRepo
 from datetime import datetime
